+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 6.8.0 (December 1, 2018)
 ### New Features
 - [#898](https://github.com/openscope/openscope/issues/898) - Add command to exit holding pattern
+- [#1003](https://github.com/openscope/openscope/issues/1003) - Add assigned runway to the aircrafts' strips